+++ src/assets/factory/insert.js
@@ -1,9 +1,4 @@
-export const getRandomIntInclusive = (min, max) => {
-  const minCeil = Math.ceil(min);
-  const maxFloor = Math.floor(max);
-
-  return Math.floor(Math.random() * (maxFloor - minCeil + 1)) + minCeil;
-};
+import {getRandomIntInclusive} from '../handler/handler';
 
 export const insert = (cards, container) => {
   let randomNumber;
