+++ src/assets/handler/handler.js
@@ -1,0 +1,6 @@
+export const getRandomIntInclusive = (min, max) => {
+  const minCeil = Math.ceil(min);
+  const maxFloor = Math.floor(max);
+
+  return Math.floor(Math.random() * (maxFloor - minCeil + 1)) + minCeil;
+};
