+++ src/main.js
@@ -1,6 +1,3 @@
-import buildFilter from './modules/filter/filter-builder';
-import buildCard from './modules/card/card-builder';
+import initApp from './modules/app-init';
 
-buildFilter();
-buildCard();
-
+initApp();
