+++ src/modules/filter/filter-builder.js
@@ -1,10 +1,7 @@
-import {filters} from '../../data/data';
 import getFilterMarkup from './filter-concreter';
 
-export const section = document.querySelector(`.main__filter`);
-
-export default () => {
-  for (const it of filters) {
+export default (data, section) => {
+  for (const it of data) {
     section.insertAdjacentHTML(
         `beforeend`, getFilterMarkup(
             {caption: it.name, amount: it.amount, state: it.state}));
